# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "研究室内の備品を管理するシステムです。",
        "title": "備品管理システム",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/assets": {
            "post": {
                "description": "資産マスター情報と個別資産情報を同時に作成します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "資産の新規作成",
                "parameters": [
                    {
                        "description": "資産作成リクエスト",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateAssetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.CreateAssetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/assets/all": {
            "get": {
                "description": "登録されている全ての個別資産情報を取得します。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "全資産の一覧取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AssetListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/assets/edit/{id}": {
            "put": {
                "description": "指定されたIDの個別資産情報を更新します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "資産情報の更新",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "資産ID (Asset ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "資産更新リクエスト",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.EditAssetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/assets/master/all": {
            "get": {
                "description": "登録されている全ての資産マスター情報を取得します。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Master"
                ],
                "summary": "全資産マスターの一覧取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AssetMasterListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/assets/master/{id}": {
            "get": {
                "description": "指定されたIDの資産マスター情報を取得します。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Master"
                ],
                "summary": "資産マスター情報の取得 (ID指定)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "資産マスターID (Asset Master ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AssetMasterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "指定されたIDの資産マスター情報を削除します。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset Master"
                ],
                "summary": "資産マスター情報の削除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "資産マスターID (Asset Master ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/assets/{id}": {
            "get": {
                "description": "指定されたIDの個別資産情報を取得します。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "資産情報の取得 (ID指定)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "資産ID (Asset ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AssetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Asset": {
            "type": "object",
            "properties": {
                "default_location": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item_master_id": {
                    "type": "integer"
                },
                "last_check_date": {
                    "type": "string"
                },
                "last_checker": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "purchase_date": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "serial_number": {
                    "type": "string"
                },
                "status_id": {
                    "type": "integer"
                }
            }
        },
        "domain.AssetMaster": {
            "type": "object",
            "properties": {
                "genre_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "management_category_id": {
                    "type": "integer"
                },
                "manufacturer": {
                    "type": "string"
                },
                "model_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "domain.CreateAssetRequest": {
            "type": "object",
            "required": [
                "quantity",
                "status_id"
            ],
            "properties": {
                "asset_master_id": {
                    "description": "--- マスター情報（任意） ---",
                    "type": "integer"
                },
                "default_location": {
                    "type": "string"
                },
                "genre_id": {
                    "type": "integer"
                },
                "last_check_date": {
                    "type": "string"
                },
                "last_checker": {
                    "type": "string"
                },
                "management_category_id": {
                    "type": "integer"
                },
                "manufacturer": {
                    "type": "string"
                },
                "model_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "purchase_date": {
                    "description": "(形式: \"YYYY-MM-DD\")",
                    "type": "string"
                },
                "quantity": {
                    "description": "--- 個別資産情報 ---",
                    "type": "integer"
                },
                "serial_number": {
                    "type": "string"
                },
                "status_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.EditAssetRequest": {
            "type": "object",
            "properties": {
                "asset_id": {
                    "description": "主キー",
                    "type": "integer"
                },
                "last_check_date": {
                    "type": "string"
                },
                "last_checker": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "purchase_date": {
                    "type": "string"
                },
                "quantity": {
                    "description": "更新時は任意",
                    "type": "integer"
                },
                "serial_number": {
                    "type": "string"
                },
                "status_id": {
                    "description": "更新時は任意, IDなのでint64",
                    "type": "integer"
                }
            }
        },
...
        "handler.AssetListResponse": {
            "type": "object",
            "properties": {
                "assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Asset"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Assets fetched successfully"
                }
            }
        },
        "handler.AssetMasterListResponse": {
            "type": "object",
            "properties": {
                "masters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.AssetMaster"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Asset masters fetched successfully"
                }
            }
        },
        "handler.AssetMasterResponse": {
            "type": "object",
            "properties": {
                "master": {
                    "$ref": "#/definitions/domain.AssetMaster"
                },
                "message": {
                    "type": "string",
                    "example": "Asset master fetched successfully"
                }
            }
        },
        "handler.AssetResponse": {
            "type": "object",
            "properties": {
                "asset": {
                    "$ref": "#/definitions/domain.Asset"
                },
                "message": {
                    "type": "string",
                    "example": "Asset fetched successfully"
                }
            }
        },
        "handler.CreateAssetResponse": {
            "type": "object",
            "properties": {
                "asset_master_id": {
                    "type": "integer",
                    "example": 101
                },
                "message": {
                    "type": "string",
                    "example": "Asset created successfully"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "不正なリクエストです"
                }
            }
        },
...
        "handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "処理に成功しました"
                }
            }
        }
    }
}